requirements.txt
# ===============================
# FastAPI CRM Backend (STABLE)
# Python 3.10 / 3.11 ONLY
# ===============================

# Core Framework
fastapi>=0.109,<0.111
uvicorn[standard]>=0.27,<0.29
python-multipart>=0.0.6

# Database (MongoDB)
motor>=3.3,<3.5
pymongo>=4.6,<4.7

# Authentication
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt>=4.1,<4.2

# Validation
pydantic[email]>=2.5,<2.7
email-validator>=2.1

# AWS S3
boto3>=1.34,<1.35
botocore>=1.34,<1.35

# Email
aiosmtplib>=3.0,<3.1

# PDF
reportlab>=4.0,<4.1

# ===============================
# Machine Learning (SAFE)
# ===============================
numpy==1.26.4
scikit-learn==1.3.2

# Environment
python-dotenv>=1.0

# Security
itsdangerous>=2.1

# Date/Time
python-dateutil>=2.8

# HTTP
httpx>=0.26,<0.27
aiohttp>=3.9,<3.10

# Utils
orjson>=3.9,<4
